HTTP 4XX responses.

+++ djangular/views/crud.py
@@ -31,15 +31,23 @@ class NgCRUDView(FormView):
         * $save - ng_save
         * $delete and $remove - ng_delete
         """
-        if request.method == 'GET':
-            if 'pk' in request.GET or self.slug_field in request.GET:
-                return self.ng_get(request, *args, **kwargs)
-            return self.ng_query(request, *args, **kwargs)
-        elif request.method == 'POST':
-            return self.ng_save(request, *args, **kwargs)
-        elif request.method == 'DELETE':
-            return self.ng_delete(request, *args, **kwargs)
-        raise ValueError('This view can not handle method %s' % request.method)
+        try:
+            if request.method == 'GET':
+                if 'pk' in request.GET or self.slug_field in request.GET:
+                    return self.ng_get(request, *args, **kwargs)
+                return self.ng_query(request, *args, **kwargs)
+            elif request.method == 'POST':
+                return self.ng_save(request, *args, **kwargs)
+            elif request.method == 'DELETE':
+                return self.ng_delete(request, *args, **kwargs)
+        except self.model_class.DoesNotExist as e:
+            return self.error_json_response(str(e), 404)
+        except ValueError as e:
+            return self.error_json_response(str(e))
+        except ValidationError as e:
+            return self.error_json_response(e.message)
+
+        return self.error_json_response('This view can not handle method {0}'.format(request.method), 405)
 
     def get_form_class(self):
         """
@@ -49,6 +57,15 @@ class NgCRUDView(FormView):
 
     def build_json_response(self, data):
         response = HttpResponse(self.serialize_to_json(data), self.content_type)
+        response['Cache-Control'] = 'no-cache'
+        return response
+
+    def error_json_response(self, message, status_code=400, detail=None):
+        response_data = {
+            "message": message,
+            "detail": detail,
+        }
+        response = HttpResponse(json.dumps(response_data, cls=DjangoJSONEncoder, separators=(',', ':')), self.content_type, status=status_code)
         response['Cache-Control'] = 'no-cache'
         return response
 
@@ -71,8 +88,8 @@ class NgCRUDView(FormView):
             object_data.append(obj['fields'])
 
         if is_queryset:
-            return json.dumps(object_data, cls=DjangoJSONEncoder)
-        return json.dumps(object_data[0], cls=DjangoJSONEncoder)  # If there's only one object
+            return json.dumps(object_data, cls=DjangoJSONEncoder, separators=(',', ':'))
+        return json.dumps(object_data[0], cls=DjangoJSONEncoder, separators=(',', ':'))  # If there's only one object
 
     def get_form_kwargs(self):
         kwargs = super(NgCRUDView, self).get_form_kwargs()
@@ -120,12 +137,19 @@ class NgCRUDView(FormView):
         if form.is_valid():
             obj = form.save()
             return self.build_json_response(obj)
-        raise ValidationError("Form not valid")
+
+        # Do not fall back to dispatch error handling, instead provide field details.
+        error_detail = {"non_field_errors": form.non_field_errors()}
+        error_detail.update(form.errors)
+        return self.error_json_response("Form not valid", detail=error_detail)
 
     def ng_delete(self, request, *args, **kwargs):
         """
         Delete object and return it's data in JSON encoding
         """
+        if 'pk' not in request.GET:
+            raise ValueError("Object id is required to delete.")
+
         obj = self.get_object()
         obj.delete()
         return self.build_json_response(obj)
